Train on 500k batches of size 2M
Apparently this config was originally designed for training on
256 GPUs. I'm training on 64, so to match the behavior at 256
GPUs (in particular the global batch size, which is a function
of the number of workers) we need to increase the gradient
accumulation to 4 (=256/64). The batch size is then
(seq_len)*(local_batch_size)*(num_workers)*(grad_accum)
= 4096*2*64*4 = 2097152.
Training for 500k steps then results in just over 1T tokens.